Calculate Eigenvalues and Eigenvectors

This helper function returns the eigenvalues in lambda and the left and right eigenvectors in forwards and backwards

Usage

eigen_system(mat)

Arguments

mat

(matrix) a square matrix

Value

a list with three elements:

lambda: eigenvalues

forwards: left eigenvector

backwards: right eigenvector

Examples

mat_T0 <- rbind(c(1-0.01,0.01,0),
               c(0.01,1-0.02,0.01),
               c(0,0.01,1-0.01))
eigen_system(mat_T0)
#> $lambda
#> [1] 1.00 0.99 0.97
#> 
#> $forwards
#>            [,1]          [,2]       [,3]
#> [1,]  0.3333333  3.333333e-01  0.3333333
#> [2,] -0.3333333 -2.796117e-16  0.3333333
#> [3,] -0.2222222  4.444444e-01 -0.2222222
#> 
#> $backwards
#>      [,1]          [,2]  [,3]
#> [1,]    1 -1.500000e+00 -0.75
#> [2,]    1 -1.040834e-15  1.50
#> [3,]    1  1.500000e+00 -0.75
#>