Forward-backward algorithm to calculate the posterior probabilities of hidden states.

Forward-backward algorithm to calculate the posterior probabilities of hidden states.

posterior_probabilities_normal(data, pi, mat_T, means, sdev)

Arguments

data

(numeric) normal data

pi

(numeric) prior probability of states

mat_T

(matrix) transition probability matrix

means

(numeric) vector with prior means

sdev

(numeric) prior standard deviation

Value

List with posterior probabilities (TO BE CORRECTED)

Details

Here details on how the calculation is made

References

Here some references

Examples

prior_mat <- rbind(c(1-0.05, 0.05, 0),
                  c(0.05, 1-0.1, 0.05),
                  c(0, 0.05, 1-0.05))

prior_means <- c(-0.1, 0.0, 0.1)
prior_sd  <- sqrt(0.1)
L <- 100

# Simulate HMM model based on normal data based on prior information
sim_data_normal <- hmm_simulate_normal_data(L = L,
                                            mat_T = prior_mat,
                                            means = prior_means,
                                            sigma = prior_sd)
pi <- sim_data_normal$pi
# pi <- get_pi(prior_mat)
hmm_norm_data <- sim_data_normal$data

# Calculate posterior probabilities of hidden states
post_prob <-  posterior_probabilities_normal(data = hmm_norm_data,
                                             pi = pi,
                                             mat_T = prior_mat,
                                             means = prior_means,
                                             sdev = prior_sd)