Converts MCMC samples into ggmcmc format

This helper function converts MCMC samples into ggmcmc format

Usage

convert_to_ggmcmc(
  x,
  pattern = c("mean", "sigma", "beta", "alpha", "pois_means", "T"),
  include_warmup = FALSE
)

Arguments

x

(mcmc_hmm_*) MCMC HMM object

pattern

(character) pattern(s) with model parameters to be included in the output

include_warmup

(logical) include warmup samples. By default FALSE

Value

data.frame compatible with functions from the ggmcmc package

Details

By default, for a given model, all parameters are converted into ggmcmc format.

The parameter pattern can be used to extract specific parameters. For instance pattern="mean" extracts all mean parameters from a hmm_mcmc_normal model.

If a specific parameter is of interest it can be matched by an exact name: pattern=c("mean[1]", "T[1,1]").

Examples

# Convert all parameters (Normal model)
convert_normal_all <- convert_to_ggmcmc(example_hmm_mcmc_normal)
unique(convert_normal_all$Parameter)
#>  [1] "mean[1]"  "mean[2]"  "mean[3]"  "sigma[1]" "T[1,1]"   "T[2,1]"  
#>  [7] "T[3,1]"   "T[1,2]"   "T[2,2]"   "T[3,2]"   "T[1,3]"   "T[2,3]"  
#> [13] "T[3,3]"  
head(convert_normal_all)
#>   Iteration Chain Parameter     value
#> 1       801     1   mean[1] 0.5448143
#> 2       802     1   mean[1] 0.5512685
#> 3       803     1   mean[1] 0.5425334
#> 4       804     1   mean[1] 0.5469280
#> 5       805     1   mean[1] 0.5406991
#> 6       806     1   mean[1] 0.5442431
tail(convert_normal_all)
#>       Iteration Chain Parameter     value
#> 15595      1995     1    T[3,3] 0.6422738
#> 15596      1996     1    T[3,3] 0.6636045
#> 15597      1997     1    T[3,3] 0.6447341
#> 15598      1998     1    T[3,3] 0.6448492
#> 15599      1999     1    T[3,3] 0.6238971
#> 15600      2000     1    T[3,3] 0.6391355

# Convert only means (Normal model)
convert_normal_means <- convert_to_ggmcmc(example_hmm_mcmc_normal, 
                                          pattern = "mean")
unique(convert_normal_means$Parameter)
#> [1] "mean[1]" "mean[2]" "mean[3]"

# Convert selected parameter (Normal model)
pattern_normal <- c("mean[1]", "sigma[1]", "T[1,1]")
convert_normal_param <- convert_to_ggmcmc(example_hmm_mcmc_normal, 
                                          pattern = pattern_normal)
unique(convert_normal_param$Parameter)
#> [1] "mean[1]"  "sigma[1]" "T[1,1]"  

# Convert all parameters (Poisson-Gamma model)
convert_pois_gamma_all <- convert_to_ggmcmc(example_hmm_mcmc_gamma_poisson)
unique(convert_pois_gamma_all$Parameter)
#>  [1] "means[1]" "means[2]" "means[3]" "beta[1]"  "beta[2]"  "beta[3]" 
#>  [7] "alpha[1]" "T[1,1]"   "T[2,1]"   "T[3,1]"   "T[1,2]"   "T[2,2]"  
#> [13] "T[3,2]"   "T[1,3]"   "T[2,3]"   "T[3,3]"